separating logic with hooks
