+++ src/useInput.js
@@ -1,0 +1,15 @@
+import { useState } from "react";
+
+const useInput = () => {
+  const [input, setInput] = useState("");
+
+  /* handle user input */
+  const handleInput = (value) => {
+    const newInput = value;
+    setInput(newInput);
+  };
+
+  return {input, setInput, handleInput};
+}
+
+export default useInput;
